:recycle: : moves settings feature to its own package
