+++ src/main/java/io/codeka/gaia/runner/StackRunner.java
@@ -3,8 +3,8 @@ package io.codeka.gaia.runner;
 import com.spotify.docker.client.DockerClient;
 import com.spotify.docker.client.exceptions.DockerException;
 import com.spotify.docker.client.messages.ContainerConfig;
-import io.codeka.gaia.bo.*;
 import io.codeka.gaia.modules.bo.TerraformModule;
+import io.codeka.gaia.settings.bo.Settings;
 import io.codeka.gaia.stacks.repository.JobRepository;
 import io.codeka.gaia.stacks.repository.StackRepository;
 import io.codeka.gaia.stacks.bo.Job;
