+++ src/test/java/io/codeka/gaia/runner/StackRunnerTest.java
@@ -5,8 +5,8 @@ import com.spotify.docker.client.exceptions.DockerException;
 import com.spotify.docker.client.messages.ContainerConfig;
 import com.spotify.docker.client.messages.ContainerCreation;
 import com.spotify.docker.client.messages.ContainerExit;
-import io.codeka.gaia.bo.*;
 import io.codeka.gaia.modules.bo.TerraformModule;
+import io.codeka.gaia.settings.bo.Settings;
 import io.codeka.gaia.stacks.repository.JobRepository;
 import io.codeka.gaia.stacks.repository.StackRepository;
 import io.codeka.gaia.stacks.bo.*;
